Refactored dynamic primitives logic into @paima/db
