+++ packages/node-sdk/db/src/mod.ts
@@ -50,5 +50,4 @@ export * from "./register-events.ts";
 export * from "./pg-connection.ts";
 export * from "./scheduled-constructors.ts";
 
-export * from "./ivm/erc20-ivm.ts";
-export * from "./ivm/erc721-ivm.ts";
+export { createDynamicTables, getPrimitivePrefix } from "./dynamic-tables.ts";
