+++ packages/node-sdk/sync/src/mod.ts
@@ -1,2 +1,3 @@
 export * from "./sync-protocols/mod.ts";
 export * from "./syncProtocolFactory.ts";
+export type { AllSyncProtocols } from "./sync-protocols/types.ts";
